requirements.txt
asgiref==3.6.0
asteval==0.9.29
attrs==22.2.0
backports.zoneinfo==0.2.1
boto3==1.26.74
botocore==1.29.74
cchardet==2.1.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.0.6
django-environ==0.9.0
django-storages==1.13.2
djangorestframework==3.14.0
drf-yasg==1.21.5
emmet-core==0.44.0
exceptiongroup==1.1.0
fabio==2022.12.1
fonttools==4.38.0
future==0.18.3
greenlet==2.0.2
h11==0.14.0
h5py==3.8.0
hdf5plugin==4.1.1
idna==3.4
imageio==2.25.1
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
latexcodec==2.0.1
lmfit==1.1.0
MarkupSafe==2.1.2
matplotlib==3.7.0
monty==2022.9.9
mp-api==0.30.10
mpmath==1.2.1
msgpack==1.0.4
networkx==3.0
numdifftools==0.9.41
numexpr==2.8.4
numpy==1.24.2
packaging==23.0
palettable==3.3.0
pandas==1.5.3
PeakUtils==1.3.4
Pillow==9.4.0
plotly==5.13.0
pluggy==1.0.0
psutil==5.9.4
psycopg2-binary==2.9.3
pybtex==0.24.0
PyCifRW==4.4.5
pydantic==1.10.5
pyfai==2023.1.0
pymatgen==2023.1.30
pyparsing==3.0.9
pyshortcuts==1.8.3
pyshorteners==1.0.1
pytest==7.2.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7.1
PyWavelets==1.4.1

PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
silx==1.1.2
six==1.16.0
spglib==2.0.2
SQLAlchemy==2.0.4
sqlparse==0.4.3
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.1
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.2.3
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
tzdata==2022.7
uncertainties==3.1.7
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
whitenoise==6.2.0
xraydb==4.5.0
xraylarch==0.9.66
zipp==3.14.0
